promote Coeff outside
