+++ src/core/structsos/__init__.py
@@ -1,9 +1,8 @@
 from .structsos import StructuralSOS
-from .utils import Coeff, radsimp, quadratic_weighting, zip_longest
+from .utils import radsimp, quadratic_weighting, zip_longest
 from .solution import SolutionStructural, SolutionStructuralSimple
 
 __all__ = [
-    'StructuralSOS',
-    'Coeff', 'radsimp', 'quadratic_weighting', 'zip_longest',
+    'StructuralSOS', 'radsimp', 'quadratic_weighting', 'zip_longest',
     'SolutionStructural', 'SolutionStructuralSimple'
 ]
