+++ src/core/structsos/quarternary/__init__.py
@@ -1,0 +1,3 @@
+from .solver import structural_sos_4vars
+
+__all__ = ['structural_sos_4vars']
