+++ src/utils/__init__.py
@@ -11,11 +11,9 @@ from .polytools import (
 )
 
 from .expression import (
-    CyclicExpr, CyclicSum, CyclicProduct, SymmetricSum, SymmetricProduct, is_cyclic_expr,
-    poly_get_factor_form, poly_get_standard_form,
-    latex_coeffs,
-    Solution, SolutionSimple,
-    congruence,
+    Coeff, CyclicExpr, CyclicSum, CyclicProduct, SymmetricSum, SymmetricProduct, is_cyclic_expr,
+    poly_get_factor_form, poly_get_standard_form, latex_coeffs,
+    Solution, SolutionSimple, congruence,
 )
 
 from .roots import (
@@ -35,11 +33,9 @@ __all__ = [
     'preprocess_text', 'pl', 'PolyReader',
     'deg', 'verify_hom_cyclic', 'verify_is_symmetric', 
     'monom_of', 'convex_hull_poly',
-    'CyclicExpr', 'CyclicSum', 'CyclicProduct', 'SymmetricSum', 'SymmetricProduct', 'is_cyclic_expr',
-    'poly_get_factor_form', 'poly_get_standard_form',
-    'latex_coeffs',
-    'Solution', 'SolutionSimple',
-    'congruence',
+    'Coeff', 'CyclicExpr', 'CyclicSum', 'CyclicProduct', 'SymmetricSum', 'SymmetricProduct', 'is_cyclic_expr',
+    'poly_get_factor_form', 'poly_get_standard_form', 'latex_coeffs',
+    'Solution', 'SolutionSimple', 'congruence',
     'RootsInfo', 'Root', 'RootAlgebraic', 'RootRational', 'RootTernary', 'RootAlgebraicTernary', 'RootRationalTernary',
     'GridRender',
     'findroot', 'find_nearest_root', 'findroot_resultant',
