+++ src/utils/expression/__init__.py
@@ -1,11 +1,12 @@
+from .coeff import Coeff
+
 from .cyclic import (
     CyclicExpr, CyclicSum, CyclicProduct, SymmetricSum, SymmetricProduct,
     is_cyclic_expr
 )
 
 from .form import (
-    poly_get_factor_form, poly_get_standard_form,
-    latex_coeffs
+    poly_get_factor_form, poly_get_standard_form, latex_coeffs
 )
 
 from .solution import (
@@ -14,9 +15,9 @@ from .solution import (
 )
 
 __all__ = [
+    'Coeff',
     'CyclicExpr', 'CyclicSum', 'CyclicProduct', 'SymmetricSum', 'SymmetricProduct', 'is_cyclic_expr',
-    'poly_get_factor_form', 'poly_get_standard_form',
-    'latex_coeffs',
+    'poly_get_factor_form', 'poly_get_standard_form', 'latex_coeffs',
     'Solution', 'SolutionSimple',
     'congruence',
 ]
